pyproject.toml
[project]
name = "text-sql-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "patlib>=0.3.7",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
